dry up map closure via macro
