+++ rust/cloud-storage/soup/src/outbound/pg_soup_repo.rs
@@ -130,3 +130,85 @@ fn type_err<E: std::fmt::Display>(e: E) -> sqlx::Error {
         type_name: e.to_string(),
     }
 }
+
+/// this defines a macro which maps the soup query types for statically checked soup queries
+/// This must be a macro because compile time queries cannot have a named type so we can't use a function
+#[macro_export]
+macro_rules! map_soup_type {
+    () => {
+        |r| match r.item_type.as_ref() {
+            "document" => Ok(::models_soup::item::SoupItem::Document(
+                ::models_soup::document::SoupDocument {
+                    id: Uuid::parse_str(&r.id).map_err(type_err)?,
+                    document_version_id: r
+                        .document_version_id
+                        .ok_or_else(|| type_err("document version id must exist"))
+                        .and_then(|s| FromStr::from_str(&s).map_err(type_err))?,
+                    owner_id: MacroUserIdStr::parse_from_str(&r.user_id)
+                        .map_err(type_err)?
+                        .into_owned(),
+                    name: r.name,
+                    file_type: r.file_type,
+                    sha: r.sha,
+                    project_id: r
+                        .project_id
+                        .as_deref()
+                        .map(Uuid::parse_str)
+                        .transpose()
+                        .map_err(type_err)?,
+                    branched_from_id: r
+                        .branched_from_id
+                        .as_deref()
+                        .map(Uuid::parse_str)
+                        .transpose()
+                        .map_err(type_err)?,
+                    branched_from_version_id: r.branched_from_version_id,
+                    document_family_id: r.document_family_id,
+                    created_at: r.created_at,
+                    updated_at: r.updated_at,
+                    viewed_at: r.viewed_at,
+                },
+            )),
+            "chat" => Ok(::models_soup::item::SoupItem::Chat(
+                ::models_soup::chat::SoupChat {
+                    id: Uuid::parse_str(&r.id).map_err(type_err)?,
+                    name: r.name,
+                    owner_id: MacroUserIdStr::parse_from_str(&r.user_id)
+                        .map_err(type_err)?
+                        .into_owned(),
+                    project_id: r
+                        .project_id
+                        .as_deref()
+                        .map(Uuid::parse_str)
+                        .transpose()
+                        .map_err(type_err)?,
+                    is_persistent: r.is_persistent.unwrap_or_default(),
+                    created_at: r.created_at,
+                    updated_at: r.updated_at,
+                    viewed_at: r.viewed_at,
+                },
+            )),
+            "project" => Ok(::models_soup::item::SoupItem::Project(
+                ::models_soup::project::SoupProject {
+                    id: Uuid::parse_str(&r.id).map_err(type_err)?,
+                    name: r.name,
+                    owner_id: MacroUserIdStr::parse_from_str(&r.user_id)
+                        .map_err(type_err)?
+                        .into_owned(),
+                    parent_id: r
+                        .project_id
+                        .as_deref()
+                        .map(Uuid::parse_str)
+                        .transpose()
+                        .map_err(type_err)?,
+                    created_at: r.created_at,
+                    updated_at: r.updated_at,
+                    viewed_at: r.viewed_at,
+                },
+            )),
+            _ => Err(sqlx::Error::TypeNotFound {
+                type_name: r.item_type,
+            }),
+        }
+    };
+}
